Move useAutocorrectingInterval
