+++ lib/AutocorrectingInterval.ts
@@ -1,3 +1,6 @@
+import { useEffect } from "react"
+import { useEffectEvent } from "./utils/UseEffectEvent"
+
 export type AutocorrectingInterval = number
 export type TimeoutFunction = (
   callback: () => void,
@@ -50,4 +53,20 @@ export const clearAutocorrectingInterval = (
 ) => {
   clearTimeout(intervals.get(interval))
   intervals.delete(interval)
+}
+
+/**
+ * A hook that runs the callback on an interval of length `intervalMillis`.
+ */
+export const useAutocorrectingInterval = (
+  callback: () => void,
+  intervalMillis: number
+) => {
+  const callbackEvent = useEffectEvent(callback)
+  useEffect(() => {
+    const interval = setAutocorrectingInterval(() => {
+      callbackEvent()
+    }, intervalMillis)
+    return () => clearAutocorrectingInterval(interval)
+  }, [intervalMillis, callbackEvent])
 }
